fix ratio off by fractions
due to multiple conversions and integer casting, precision gets lost,
thus an image may be requested with a slightly incorrect ratio and the
resulting width and height might still be too large. Fix this by running
it 10 times and assign original width and height at the end of it.

fix precision anytime width or height is bigger than original image.